+++ packages/sv-utils/api-surface.md
@@ -742,15 +742,6 @@ type Package = {
 	keywords?: string[];
 	workspaces?: string[];
 };
-declare const commonFilePaths: {
-	readonly packageJson: 'package.json';
-	readonly svelteConfig: 'svelte.config.js';
-	readonly svelteConfigTS: 'svelte.config.ts';
-	readonly jsconfig: 'jsconfig.json';
-	readonly tsconfig: 'tsconfig.json';
-	readonly viteConfig: 'vite.config.js';
-	readonly viteConfigTS: 'vite.config.ts';
-};
 declare function fileExists(cwd: string, filePath: string): boolean;
 
 declare function loadFile(cwd: string, filePath: string): string;
@@ -761,18 +752,6 @@ declare function loadPackageJson(cwd: string): {
 	data: Package;
 	generateCode: () => string;
 };
-/**
- * @deprecated Use {@link loadFile} instead. This alias will be removed in a future version.
- */
-declare const readFile: typeof loadFile;
-/**
- * @deprecated Use {@link saveFile} instead. This alias will be removed in a future version.
- */
-declare const writeFile: typeof saveFile;
-/**
- * @deprecated Use {@link loadPackageJson} instead. This alias will be removed in a future version.
- */
-declare const getPackageJson: typeof loadPackageJson;
 type ColorInput = string | string[];
 declare const color: {
 	addon: (str: ColorInput) => string;
@@ -810,7 +789,6 @@ export {
 	index_d_exports as Walker,
 	type YamlDocument,
 	color,
-	commonFilePaths,
 	constructCommand,
 	createPrinter,
 	index_d_exports$1 as css,
@@ -818,7 +796,6 @@ export {
 	detect,
 	downloadJson,
 	fileExists,
-	getPackageJson,
 	index_d_exports$2 as html,
 	isVersionUnsupportedBelow,
 	index_d_exports$3 as js,
@@ -827,7 +804,6 @@ export {
 	loadPackageJson,
 	parse,
 	pnpm_d_exports as pnpm,
-	readFile,
 	resolveCommand,
 	resolveCommandArray,
 	sanitizeName,
@@ -835,7 +811,6 @@ export {
 	splitVersion,
 	index_d_exports$4 as svelte,
 	text_d_exports as text,
-	transforms,
-	writeFile
+	transforms
 };
 ```
