+++ packages/sv/src/core/common.ts
@@ -324,3 +324,13 @@ export function updateAgent(
 		fs.writeFileSync(agentPath, content);
 	}
 }
+
+export const commonFilePaths = {
+	packageJson: 'package.json',
+	svelteConfig: 'svelte.config.js',
+	svelteConfigTS: 'svelte.config.ts',
+	jsconfig: 'jsconfig.json',
+	tsconfig: 'tsconfig.json',
+	viteConfig: 'vite.config.js',
+	viteConfigTS: 'vite.config.ts'
+} as const;
