+++ packages/sv/src/core/workspace.ts
@@ -3,13 +3,13 @@ import {
 	type AstTypes,
 	js,
 	parse,
-	commonFilePaths,
 	loadFile,
 	loadPackageJson
 } from '@sveltejs/sv-utils';
 import * as find from 'empathic/find';
 import fs from 'node:fs';
 import path from 'node:path';
+import { commonFilePaths } from './common.ts';
 import type { OptionDefinition, OptionValues } from './options.ts';
 import { detectPackageManager } from './package-manager.ts';
